finished assignment but member can't make a new post

+++ src/db/queries.posts.js
@@ -1,15 +1,22 @@
 const Post = require("./models").Post;
 const Topic = require("./models").Topic;
+const Authorizer = require("../policies/post");
 
 module.exports = {
   addPost(newPost, callback) {
-    return Post.create(newPost)
-      .then(post => {
-        callback(null, post);
-      })
-      .catch(err => {
-        callback(err);
-      });
+    const authorized = new Authorizer(req.user).new();
+    if (authorized) {
+      return Post.create(newPost)
+        .then(post => {
+          callback(null, post);
+        })
+        .catch(err => {
+          callback(err);
+        });
+    } else {
+      req.flash("notice", "You are not authorized to do that.");
+      callback(401);
+    }
   },
   getPost(id, callback) {
     return Post.findById(id)
@@ -20,33 +27,45 @@ module.exports = {
         callback(err);
       });
   },
-  deletePost(id, callback) {
-    return Post.destroy({
-      where: { id }
-    })
-      .then(deletedRecordsCount => {
-        callback(null, deletedRecordsCount);
+  deletePost(req, callback) {
+    const authorized = new Authorizer(req.user).destroy();
+    if (authorized) {
+      return Post.destroy({
+        where: { id }
       })
-      .catch(err => {
-        callback(err);
-      });
-  },
-  updatePost(id, updatedPost, callback) {
-    return Post.findById(id).then(post => {
-      if (!post) {
-        return callback("Post not found");
-      }
-
-      post
-        .update(updatedPost, {
-          fields: Object.keys(updatedPost)
-        })
-        .then(() => {
-          callback(null, post);
+        .then(deletedRecordsCount => {
+          callback(null, deletedRecordsCount);
         })
         .catch(err => {
           callback(err);
         });
-    });
+    } else {
+      req.flash("notice", "You are not authorized to do that.");
+      callback(401);
+    }
+  },
+  updatePost(req, updatedPost, callback) {
+    const authorized = new Authorizer(req.user).update();
+    if (authorized) {
+      return Post.findById(id).then(post => {
+        if (!post) {
+          return callback("Post not found");
+        }
+
+        post
+          .update(updatedPost, {
+            fields: Object.keys(updatedPost)
+          })
+          .then(() => {
+            callback(null, post);
+          })
+          .catch(err => {
+            callback(err);
+          });
+      });
+    } else {
+      req.flash("notice", "You are not authorized to do that.");
+      callback(401);
+    }
   }
 };
